remove teeReader
