+++ common/utils/lowhttp/lowhttp2/dumper.go
@@ -9,6 +9,7 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
+	"github.com/yaklang/yaklang/common/utils"
 	"io"
 	"net"
 	"net/http"
@@ -129,7 +130,7 @@ func DumpRequestOut(req *http.Request, body bool) ([]byte, error) {
 	quitReadCh := make(chan struct{})
 	// Wait for the request before replying with a dummy response:
 	go func() {
-		req, err := http.ReadRequest(bufio.NewReader(pr))
+		req, err := utils.ReadHTTPRequestFromReader(bufio.NewReader(pr))
 		if err == nil {
 			// Ensure all the body is read; otherwise
 			// we'll get a partial dump.
